Fix variable reference in init.sh
It appears that the is_process_service function defined in init.sh is improperly using the globally defined `$SERVICE` variable when identifying whether a given PID corresponds to a service, instead of the local `$SERVICE_NAME` variable as intended.

Additionally, running the `ps | grep` command failed for me because the output from `ps` was being truncated on my operating system (I'm adapting this script to run on IBM AIX). Adding the `ewww` flag prevents output truncation, so the `grep` can correctly find the name of the service in the `ps` output.