Change sleep value to 5 seconds in init.sh
If a server has a particularly long startup sequence (or the JVM just takes a while to start), this script will spuriously report sucess for `init.sh start`. Our infrastructure will then asyncronously call this later and realize that the service is in fact down. This results in the service flipping on and off instead of failing fast. Bumping this value to 5 seconds does not solve the problem, but  does give services a longer window in which to fail on startup.