Disable tests which start mock service (#508)
We've encountered an issue where a change to the CI environment is causing processes to persist as zombie processes instead of properly being cleaned up. This causes our builds to fails due to timeout since go-java-launcher polls will wait for up to 4 minutes for its subprocesses to get cleaned up.

I'd like to get develop green and publishable again while we find a solution to the root cause.