Disable string deduplication
Internally we've seen that string deduplication can cause GC overheads to go way up (from 1.5% STW to ~15% STW). This may contribute to the number of people who've tried G1 and needed to turn it off. Given the all-or-nothing nature of this flag, this makes the default garbage collector be rather unsuitable for general use. We should probably make the hybrid profile not enable it by default, and instead enable it in specific places where it makes sense.