Record heap dumps when java services OOM
Previously OOMs did not result in hreap dumps, only crash files.
Now a heap dump is recorded to `var/log/heapdump.hprof` in the
event of an OOM. If an hprof already exists in this location, it
will not be overridden, and no hprof will be recorded to prevent
services that quickly OOM on startup from filling the disk.
Any time a service OOMs there's action required to figure out
precisely what went wrong. Without recording heap, there may
not be any bread crumbs to understand the failure scenario.

Risks:
Heap dumps are large, even a single heap dump may dominate disk
utilization compared to all other logging combined. In the worst
case this may fill the disk and prevent a service from restarting,
however this is a price we pay for the ability to undesrtand how
products have failed.
Services which already set `-XX:+HeapDumpOnOutOfMemoryError` and
expect multiple heap dumps to be recorded will need to add the
`-XX:HeapDumpPath=var/log/` option to override defaults if the wish
to avoid changing behavior.