Add a marker task for writing just product dependencies locks
In some circumstances, it is useful to do just one type of lock
writing at a time, without running everything that keys off of
--write-locks. This is less common for product-dependencies.lock
updates, which are now usually slower than other types, but this
still comes in useful in certain situations.

Adds a marker task named `writeProductDependenciesLocks`, matching
the convention of `writeVersionsLocks` in gradle-consistent-versions.
This modifies the behavior of the existing createManifest tasks.