sls-packaging no longer sets `-XX:+UseContainerCpuShares`
This is part of a deprecation campaign to allow jvms in containers
to use ergonomics profiles based on the host system as opposed
to requests-based overrides.
Removing `UseContainerCpuShares` will have no impact in the
short term because go-java-launcher overrides the
`ActiveProcessorCount` property based on requests by default,
unless [container support is
disabled](https://github.com/palantir/go-java-launcher#disabling-container-support).
An audit of uses of that flag shows primarily tests will be
impacted, relatively low risk that this will impact
services. In the off-chance it does make a change for services,
the result will be the behavior that we ultimately want, without
the slow rollout we intend for most products.