Hybrid (G1) GC uses the defeault 200ms MaxGCPauseMillis on JDK-21+
It seems that the heuristics have gone a bit sideways in JDK-21
causing degenerate full gc pauses unnecessarily when we configure
`-XX:MaxGCPauseMillis=500`, however setting the default 200ms value
resolves this behavior.

Note that we deployed the increase from 200ms to 500ms at a time
when container cpu shares informed the JDKs processor count, thus
gc threads/etc, which is no longer the case. As such, it should
be safe (and generally more stable) to use the hardened default
values from the jdk.